Unit tests and New button, firebase instancing overhaul
Wrote 5 unit tests using the new mock database files, added a new button with custom_button and wrote a unit test for it as well. Changed some of the code in home, scoreboard, and user profile to match new database interface which will choose between a test database and a real database depending on if you are testing or not.